requirements.txt
apispec==6.3.0
attrs==23.1.0
Babel==2.12.1
blinker==1.6.2
Bootstrap-Flask==2.2.0
click==8.1.3
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.1
flake8==6.0.0
flake8-html==0.4.3
Flask==2.3.2
Flask-Babel==2.0.0
flask-bootstrap5==0.1.dev1
Flask-CKEditor==0.4.6
flask-form==1.0.0
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
limits==3.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-sqlalchemy==0.26.1
mccabe==0.7.0
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.1
pluggy==1.2.0
prison==0.2.1
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
rich==13.4.2
six==1.16.0
SQLAlchemy==1.4.48
SQLAlchemy-Utils==0.41.1
tomli==2.0.1
typing_extensions==4.7.0
Werkzeug==2.3.6
wrapt==1.15.0
WTForms==3.0.1
